Move comment permissions classes
Move comment-related permissions
classes to comments.permissions.py
